Adds support for creating notebooks at paths

+++ noterminal/__init__.py
@@ -12,6 +12,7 @@ from notebook.base.handlers import IPythonHandler
 from pathlib import Path
 import uuid
 import aljpy
+from urllib.parse import unquote
 
 DEFAULT = """{"cells": [], "metadata": {}, "nbformat": 4, "nbformat_minor": 2}"""
 
@@ -19,30 +20,43 @@ class NoterminalHandler(IPythonHandler):
 
     @web.authenticated
     def get(self):
-        api_path = f'.{aljpy.humanhash(n=2)}.ipynb'
-        os_path = Path(self.contents_manager._get_os_path(api_path))
-        self.log.info(f'Creating noterminal at {api_path}')
-        #TODO: Do this via the contents_manager
-        os_path.write_text(DEFAULT)
-        self.redirect(f'/notebooks/{api_path}')
+        path = Path(self.request.uri)
+        # If the frontend's sent us a full ipynb path, then get the parent
+        if path.suffix != '':
+            path = path.parent
+
+        dir = path.relative_to('/noterminal-create')
+
+        api_path = dir / f'.{aljpy.humanhash(n=2)}.ipynb'
+        os_path = Path(self.contents_manager._get_os_path(str(api_path)))
+        if os_path.parent.exists():
+            self.log.info(f'Creating noterminal at {api_path}')
+            #TODO: Do this via the contents_manager
+            os_path.write_text(DEFAULT)
+            self.redirect(f'/notebooks/{api_path}')
+        else:
+            self.send_error(400)
+
 
 class DeleteHandler(IPythonHandler):
 
     @web.authenticated
     def get(self):
-        api_path = self.get_argument('path', '')
+        api_path = unquote(self.get_argument('path', ''))
         kernel = self.get_argument('kernel', '')
-        if api_path.startswith('.'):
+        if Path(api_path).name.startswith('.'):
             self.log.info(f'Removing noterminal at {api_path}')
             self.kernel_manager.shutdown_kernel(kernel)
             os_path = Path(self.contents_manager._get_os_path(api_path))
             os_path.unlink() 
+        else:
+            self.log.info('No noterminal notebook at {api_path} to remove')
 
 def load_jupyter_server_extension(nb_server_app):
     app = nb_server_app.web_app
 
-    route = url_path_join(app.settings['base_url'], '/noterminal')
+    route = url_path_join(app.settings['base_url'], '/noterminal-create/?.*')
     app.add_handlers('.*$', [(route, NoterminalHandler)])
 
-    route = url_path_join(app.settings['base_url'], '/noterminal/delete')
+    route = url_path_join(app.settings['base_url'], '/noterminal-delete.*')
     app.add_handlers('.*$', [(route, DeleteHandler)])
